3 - refactor
